fix cron to do some monthly tasks, not just daily tasks

+++ cron/gomngr.sh
@@ -49,7 +49,7 @@ if [ -e /tmp/gomngr.lock ]; then
 fi
 
 TOMORROW=`date --date="1 day 05:00:00" +%s`
-
+NEXTMONTH=`date --date="$(date +'%Y-%m-01') + 1 month 05:00:00"`
 
 while true; do
 
@@ -102,9 +102,9 @@ while true; do
   done </tmp/gomngr.list
   echo "${NBFILES} tasks handled"
 
-  if [ $NOW -gt $TOMORROW ]; then
-    echo "${NOW}: time for daily tasks"
-    TOMORROW=`date --date="1 day 05:00:00" +%s`
+  if [ $NOW -gt $NEXTMONTH ]; then
+    echo "${NOW}: time for monthly tasks"
+    NEXTMONTH=`date --date="$(date +'%Y-%m-01') + 1 month 05:00:00"`
     echo "Check for account upcoming expiration"
     /opt/crontask.sh test_expiration
     if [ $EXIT_REQUEST -eq 1 ]; then
@@ -121,7 +121,10 @@ while true; do
       echo "Exit requested"
       exit 0
     fi
+  fi
+  if [ $NOW -gt $TOMORROW ]; then
     echo "Check for reservation removal"
+    TOMORROW=`date --date="1 day 05:00:00" +%s`
     /opt/crontask.sh reservation_remove
     if [ $EXIT_REQUEST -eq 1 ]; then
       rm /tmp/gomngr.lock
